improve spot instances upgrades

+++ service/controller/azuremachinepool/handler/nodepool/create_terminate_old_workers.go
@@ -33,53 +33,66 @@ func (r *Resource) terminateOldWorkersTransition(ctx context.Context, obj interf
 		return currentState, microerror.Mask(err)
 	}
 
-	r.Logger.Debugf(ctx, "filtered instance IDs for old instances")
-	r.Logger.Debugf(ctx, "terminating %d old worker instances", len(oldInstances))
+	if len(oldInstances) > 0 {
+		r.Logger.Debugf(ctx, "There are still %d workers from the previous release running", len(oldInstances))
 
-	if azureMachinePool.Spec.Template.SpotVMOptions != nil {
-		r.Logger.Debugf(ctx, "using simulate eviction to delete instance for spot instances node pool")
+		r.Logger.Debugf(ctx, "terminating %d old worker instances", len(oldInstances))
 
-		virtualMachineScaleSetVMsClient, err := r.ClientFactory.GetVirtualMachineScaleSetVMsClient(ctx, azureMachinePool.ObjectMeta)
-		if err != nil {
-			return currentState, microerror.Mask(err)
-		}
+		if azureMachinePool.Spec.Template.SpotVMOptions != nil {
+			r.Logger.Debugf(ctx, "using simulate eviction to delete instance for spot instances node pool")
 
-		for _, i := range oldInstances {
-			_, err = virtualMachineScaleSetVMsClient.SimulateEviction(ctx, key.ClusterID(&azureMachinePool), key.NodePoolVMSSName(&azureMachinePool), *i.InstanceID)
+			virtualMachineScaleSetVMsClient, err := r.ClientFactory.GetVirtualMachineScaleSetVMsClient(ctx, azureMachinePool.ObjectMeta)
+			if err != nil {
+				return currentState, microerror.Mask(err)
+			}
+
+			for _, i := range oldInstances {
+				_, err = virtualMachineScaleSetVMsClient.SimulateEviction(ctx, key.ClusterID(&azureMachinePool), key.NodePoolVMSSName(&azureMachinePool), *i.InstanceID)
+				if err != nil {
+					return DeploymentUninitialized, microerror.Mask(err)
+				}
+			}
+
+		} else {
+			var ids compute.VirtualMachineScaleSetVMInstanceRequiredIDs
+			{
+				var strIds []string
+				for _, i := range oldInstances {
+					strIds = append(strIds, *i.InstanceID)
+				}
+
+				ids = compute.VirtualMachineScaleSetVMInstanceRequiredIDs{
+					InstanceIds: to.StringSlicePtr(strIds),
+				}
+			}
+
+			virtualMachineScaleSetsClient, err := r.ClientFactory.GetVirtualMachineScaleSetsClient(ctx, azureMachinePool.ObjectMeta)
+			if err != nil {
+				return currentState, microerror.Mask(err)
+			}
+
+			res, err := virtualMachineScaleSetsClient.DeleteInstances(ctx, key.ClusterID(&azureMachinePool), key.NodePoolVMSSName(&azureMachinePool), ids, nil)
+			if err != nil {
+				return DeploymentUninitialized, microerror.Mask(err)
+			}
+			_, err = virtualMachineScaleSetsClient.DeleteInstancesResponder(res.Response())
 			if err != nil {
 				return DeploymentUninitialized, microerror.Mask(err)
 			}
 		}
 
-	} else {
-		var ids compute.VirtualMachineScaleSetVMInstanceRequiredIDs
-		{
-			var strIds []string
-			for _, i := range oldInstances {
-				strIds = append(strIds, *i.InstanceID)
-			}
+		r.Logger.Debugf(ctx, "terminated %d old worker instances", len(oldInstances))
 
-			ids = compute.VirtualMachineScaleSetVMInstanceRequiredIDs{
-				InstanceIds: to.StringSlicePtr(strIds),
-			}
-		}
-
-		virtualMachineScaleSetsClient, err := r.ClientFactory.GetVirtualMachineScaleSetsClient(ctx, azureMachinePool.ObjectMeta)
-		if err != nil {
-			return currentState, microerror.Mask(err)
-		}
-
-		res, err := virtualMachineScaleSetsClient.DeleteInstances(ctx, key.ClusterID(&azureMachinePool), key.NodePoolVMSSName(&azureMachinePool), ids, nil)
-		if err != nil {
-			return DeploymentUninitialized, microerror.Mask(err)
-		}
-		_, err = virtualMachineScaleSetsClient.DeleteInstancesResponder(res.Response())
-		if err != nil {
-			return DeploymentUninitialized, microerror.Mask(err)
-		}
+		return currentState, nil
 	}
 
-	r.Logger.Debugf(ctx, "terminated %d old worker instances", len(oldInstances))
+	// All old nodes are terminated.
 
-	return WaitForOldWorkersToBeGone, nil
+	// Enable cluster autoscaler for this nodepool.
+	err = r.enableClusterAutoscaler(ctx, azureMachinePool)
+	if err != nil {
+		return DeploymentUninitialized, microerror.Mask(err)
+	}
+
+	return DeploymentUninitialized, nil
 }
